refactor: move llm_d config and phases
