--- config/llm_d/models.yaml
@@ -1,25 +1,0 @@
-models:
-
-  qwen3-0-6b:
-    served_model_name: Qwen/Qwen3-0.6B
-    uri: hf://Qwen/Qwen3-0.6B
-    resources:
-      requests:
-        cpu: "4"
-        memory: 16Gi
-        nvidia.com/gpu: "1"
-      limits:
-        cpu: "4"
-        memory: 16Gi
-        nvidia.com/gpu: "1"
-
-  llama-3-1-8b-instruct-fp8:
-    served_model_name: llama-3-1-8b-instruct-fp8
-    uri: oci://registry.redhat.io/rhelai1/modelcar-llama-3-1-8b-instruct-fp8-dynamic:1.5
-    resources:
-      requests:
-        cpu: "4"
-        memory: 8Gi
-        nvidia.com/gpu: "1"
-      limits:
-        nvidia.com/gpu: "1"
