--- config/llm_d/presets.yaml
@@ -1,14 +1,0 @@
-aliases:
-  cks: smoke
-
-presets:
-
-  smoke:
-    model: qwen3-0-6b
-    smoke_request: default
-    benchmark: null
-
-  benchmark-short:
-    model: llama-3-1-8b-instruct-fp8
-    smoke_request: default
-    benchmark: short
