+++ projects/llm_d/README.md
@@ -6,11 +6,20 @@ The current implementation is intentionally narrow:
 
 - target only downstream `LLMInferenceService`
 - keep the public interface compatible with current Fournos phase execution
-- use checked-in presets and manifests instead of a large mutable config surface
+- use checked-in config chunks and manifests instead of a large mutable config surface
+
+Configuration layout:
+
+- base config: [`orchestration/config.yaml`](./orchestration/config.yaml)
+- config chunks: [`orchestration/config.d`](./orchestration/config.d)
+- presets: [`orchestration/presets.d`](./orchestration/presets.d)
+- manifests: [`orchestration/manifests`](./orchestration/manifests)
 
 Main entrypoints:
 
-- CI phase wrapper: [ci.py](/Users/aperdomo/workspace/redhat/forge/projects/llm_d/orchestration/ci.py)
-- Prepare flow: [prepare_llmd.py](/Users/aperdomo/workspace/redhat/forge/projects/llm_d/orchestration/prepare_llmd.py)
-- Test flow: [test_llmd.py](/Users/aperdomo/workspace/redhat/forge/projects/llm_d/orchestration/test_llmd.py)
-- Shared runtime/config loader: [llmd_runtime.py](/Users/aperdomo/workspace/redhat/forge/projects/llm_d/orchestration/llmd_runtime.py)
+- CI phase wrapper: [`orchestration/ci.py`](./orchestration/ci.py)
+- CLI wrapper: [`orchestration/cli.py`](./orchestration/cli.py)
+- Shared runtime/config loader: [`orchestration/llmd_runtime.py`](./orchestration/llmd_runtime.py)
+- Toolbox prepare command: [`toolbox/prepare/main.py`](./toolbox/prepare/main.py)
+- Toolbox test command: [`toolbox/test/main.py`](./toolbox/test/main.py)
+- Toolbox cleanup command: [`toolbox/cleanup/main.py`](./toolbox/cleanup/main.py)
