+++ projects/llm_d/orchestration/config.d/platform.yaml
