+++ projects/llm_d/orchestration/config.d/workloads.yaml
@@ -5,7 +5,6 @@ smoke_requests:
     temperature: 0.7
 
 benchmarks:
-
   short:
     job_name: guidellm-benchmark
     image: ghcr.io/vllm-project/guidellm:v0.5.4
