+++ projects/llm_d/orchestration/presets.d/presets.yaml
@@ -1,9 +1,17 @@
 __multiple: true
 
-pvc_rwx:
-  prepare.pvc.name: storage-rwx
-  prepare.pvc.access_mode: ReadWriteMany
+smoke:
+  runtime.selected_preset: smoke
+  runtime.model_key: qwen3-0-6b
+  runtime.smoke_request_key: default
+  runtime.benchmark_key: null
 
+benchmark-short:
+  runtime.selected_preset: benchmark-short
+  runtime.model_key: llama-3-1-8b-instruct-fp8
+  runtime.smoke_request_key: default
+  runtime.benchmark_key: short
 
-llama-70b:
-  tests.llmd.inference_service.model: llama3-3-70b
+cks:
+  extends:
+    - smoke
